refactor(chat): extract Composer out of Chat.tsx
