+++ src/components/chat/index.ts
@@ -27,3 +27,4 @@ export {
   type AttachmentUploader,
   type PendingAttachment,
 } from './ComposerAttachments'
+export { Composer } from './Composer'
